Limit the length of a packet in header to 2^14 bytes.
stdin and stdout buffers are usually limited to 64KiB. The previous maximum length was 64KiB, which means that for busy TCP sockets there was almost never the opportunity of using the proxied_tcp::allocate_write() to avoid one extra buffer copy.
To keep the overhead similar the header struct has been modified to fit into a 32-bit integer, penalizing the id size (you shouldn't have thousands of sockets active anyway).

+++ src/env.hpp
@@ -24,9 +24,9 @@ void udp_front_loop(boost::asio::yield_context yield);
 
 template<bool client> void read_remote(boost::asio::yield_context yield);
 
-void send_output(opcodes opcode, uint64_t id, uint16_t len, const void* data);
-inline void send_output(opcodes opcode, uint64_t id) { send_output(opcode, id, 0, 0); }
-char* allocate_output(opcodes opcode, uint64_t id, uint16_t len);
+void send_output(opcodes opcode, uint16_t id, uint16_t len, const void* data);
+inline void send_output(opcodes opcode, uint16_t id) { send_output(opcode, id, 0, 0); }
+char* allocate_output(opcodes opcode, uint16_t id, uint16_t len);
 void commit_output();
 void abort_output(uint16_t len);
 std::tuple<uint64_t, size_t> get_output_statistics();
